Move Lingo API into Lingo directory
